Move commentTreeUtils to its own file to break an import cycle
